requirements.txt
Flask==3.1.0
mysql-connector==2.2.9
numpy==2.2.5
pandas==2.2.3
pdfkit==1.0.0
pytest==8.3.5
python-dotenv==1.0.0
scikit-learn==1.6.1
SQLAlchemy==2.0.40
Werkzeug==3.1.3
WTForms==3.2.1